Player Target Follow and shooting system on going to finish

+++ Assets/Scripts/Data/ValueObject/BulletTypesData.cs
@@ -5,7 +5,7 @@ namespace Data.ValueObject
     [Serializable]
     public class BulletTypesData
     {
-        public int Damage;
-        public int FireRate;
+        public float Damage;
+        public float FireRate;
     }
 }
